separate tools.js into toolmanager.js and discrete files for each tool
